Add CodeFights to Sites for Practice section
CodeFights is one of the very few competitive programming websites which I still frequent. It is nicely gamified and helps to solve (at least) a problem a day. I think it's worth adding it to this list.

Disclaimer: I'm not affiliated with CodeFights organization, but I won several of their t-shirts.