Updated Latest revised Edition of Steven Helim
The Latest 3rd edition (2013) is very much more detailed than the First Edition and it contains nearly 300 pages of more Amazing knowledge than the edition which is currently uploaded on this repository. The latest edition has many more examples and updated algorithms which are essential for competitive programming.
It will be beneficial for Everybody  who is coming to checkout this repository to get the latest edition and will keep it updated.